Getting the logging right
