+++ lega/conf/loggers/console.yaml
@@ -28,13 +28,15 @@ handlers:
     stream: ext://sys.stdout
 
 formatters:
-  json:
-    (): lega.conf.logging.JSONFormatter
-    format: '(asctime) (name) (process) (processName) (levelname) (lineno) (funcName) (message)'
   lega:
-    format: '[{asctime:<20}][{name}][{process:d} {processName:>15}][{levelname}] (L:{lineno}) {funcName}: {message}'
+    fmt: '[{asctime:<20}][{name}][{process:d} {processName:>15}][{levelname}] (L:{lineno}) {funcName}: {message}'
     style: '{'
     datefmt: '%Y-%m-%d %H:%M:%S'
   simple:
-    format: '[{name:^10}][{levelname:^6}] (L{lineno}) {message}'
+    fmt: '[{name:^10}][{levelname:^6}] (L{lineno}) {message}'
     style: '{'
+  json:
+    (): lega.utils.logging.JSONFormatter
+    fmt: '(correlation_id) (asctime) (name) (process) (processName) (levelname) (lineno) (funcName) (message)'
+    validate: False
+    datefmt: '%Y-%m-%d %H:%M:%S'
