+++ lega/conf/loggers/logstash-debug.yaml
@@ -5,41 +5,11 @@ root:
 
 loggers:
   lega:
-    level: DEBUG
+    level: INFO
     handlers: [logstash,console]
     propagate: True
     qualname: lega
-  amqp:
-    level: DEBUG
-    handlers: [logstash,console]
-  db:
-    level: DEBUG
-    handlers: [logstash,console]
-  crypto:
-    level: DEBUG
-    handlers: [logstash,console]
-  asyncio:
-    level: DEBUG
-    handlers: [console]
-  aiopg:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.access:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.client:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.internal:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.server:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.web:
-    level: DEBUG
-    handlers: [console]
-  aiohttp.websocket:
+  py.warnings:
     level: DEBUG
     handlers: [console]
 
@@ -55,17 +25,17 @@ handlers:
   logstash:
     class: lega.conf.logging.LEGAHandler
     formatter: json
-    host: ega_monitor
-    port: 5600
+    host: logs-logstash
+    port: 9000
 
 formatters:
-  json:
-    (): lega.conf.logging.JSONFormatter
-    format: '(asctime) (name) (process) (processName) (levelname) (lineno) (funcName) (message)'
   lega:
     format: '[{asctime:<20}][{name}][{process:d} {processName:>15}][{levelname}] (L:{lineno}) {funcName}: {message}'
     style: '{'
     datefmt: '%Y-%m-%d %H:%M:%S'
   simple:
+    format: '[{correlation_id}][{name:^10}][{levelname:^6}] (L{lineno}) {message}'
+    style: '{'
+  simple_no_correlation:
     format: '[{name:^10}][{levelname:^6}] (L{lineno}) {message}'
     style: '{'
