+++ lega/conf/loggers/vector.yaml
@@ -6,9 +6,12 @@ root:
 loggers:
   lega:
     level: DEBUG
-    handlers: [logstash,console]
+    handlers: [udp,console]
     propagate: True
     qualname: lega
+  py.warnings:
+    level: DEBUG
+    handlers: [console]
 
 
 handlers:
@@ -17,21 +20,21 @@ handlers:
     level: NOTSET
   console:
     class: logging.StreamHandler
-    formatter: simple
+    formatter: lega
     stream: ext://sys.stderr
-  logstash:
-    class: lega.utils.logging.LEGAHandler
+  udp:
+    class: logging.handlers.DatagramHandler
     formatter: json
-    host: ega_monitor
-    port: 5600
+    host: logs
+    port: 9000
 
 formatters:
   lega:
-    format: '[{asctime:<20}][{name}][{process:d} {processName:>15}][{levelname}] (L:{lineno}) {funcName}: {message}'
+    fmt: '[{asctime:<20}][{name}][{process:d} {processName:>15}][{levelname}] (L:{lineno}) {funcName}: {message}'
     style: '{'
     datefmt: '%Y-%m-%d %H:%M:%S'
   simple:
-    format: '[{name:^10}][{levelname:^6}] (L{lineno}) {message}'
+    fmt: '[{correlation_id}][{name:^10}][{levelname:^6}] (L{lineno}) {message}'
     style: '{'
   json:
     (): lega.utils.logging.JSONFormatter
